requirements.txt
asgiref==3.8.1
Django==5.1.2
django-tinymce==4.1.0
psycopg2==2.9.10
sqlparse==0.5.1
tzdata==2024.2
